projectLourdes/requirements.txt
scikit-learn==0.24.2
flask==1.1.2
xgboost==1.5.1
matplotlib==3.4.3
seaborn==0.11.2
panda==0.3.1
numpy==1.20.3
werkzeug==2.0.2
openpyxl==3.0.10
